pyproject.toml
[project]
name = "shieldai"
version = "0.1.0"
description = "SHIELD AI - Real-time industrial discharge anomaly detection"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.26.0",
    "pandas>=2.0.0,<3.0.0",
    "pathway>=0.29.1",
    "streamlit>=1.43.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "fpdf2>=2.8.0",
    "scikit-learn>=1.6.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
]
